+++ app/ospf/components_core.js
@@ -6,4 +6,5 @@ var coreComponents = [
     "core/Label",
     "core/DbManager",
     "core/CheckBox",
+    "core/EventManager",
 ];