+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # 1.6
 - Refactored component creation. Component are now created with:
-`let myComponent = await ComponentClass.Create();`
+```javascript
+let myComponent = await ComponentClass.Create({
+    property1: 123,
+    property2: "abc"
+});
+```
+
+- Properties are assigned automatically to the component.
+
 - `init()` has been replaced with `onAfterInit` in custom components:
 ```javascript
 class HelloWorld extends Component {
@@ -8,7 +16,8 @@ class HelloWorld extends Component {
         this.addNameClickCallback = emptyCallback;
         ...
 ```
-It is no longer needed to call `super.init()` inside this new method.
+
+- It is no longer needed to call `super.init()` inside this new method.
 
 # 1.5
 created changelog